Adicionados scripts para rodar o exemplo em modo home ou em modo rotas
