+++ example/.vscode/launch.json
@@ -1,0 +1,24 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "Routes",
+            "request": "launch",
+            "type": "dart",
+            "args": [
+                "--target=lib/main_routes.dart"
+            ]
+        },
+        {
+            "name": "Home",
+            "request": "launch",
+            "type": "dart",
+            "args": [
+                "--target=lib/main_home.dart"
+            ]
+        }
+    ]
+}
