+++ example/lib/main_home.dart
@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:localization/localization.dart';
 
-void main() async {
-  runApp(MyApp());
-}
+void main() => runApp(MyApp());
 
 class MyApp extends StatelessWidget {
   // This widget is the root of your application.
